refactor: fix import path for PasswordRecoveryPage component
